% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrainsReg.R
\name{cglasso}
\alias{cglasso}
\title{Fits regularization paths for group-lasso penalized learning problems with linear constraints.}
\usage{
cglasso(y, Z, Zc = NULL, k, W = rep(1, times = p), intercept = TRUE,
  A = kronecker(matrix(1, ncol = p), diag(k)), b = rep(0, times = k),
  u = 1, mu_ratio = 1.01, lam = NULL, nlam = 100,
  lambda.factor = ifelse(n < p1, 0.05, 0.001), dfmax = p,
  pfmax = min(dfmax * 1.5, p), tol = 0, outer_maxiter = 3e+08,
  outer_eps = 1e-08, inner_maxiter = 1e+06, inner_eps = 1e-08)
}
\arguments{
\item{y}{a vector of response variable with length \eqn{n}.}

\item{Z}{a matrix, of dimension \eqn{n \times p1}{n*p1}, for predictors to be imposed with group lasso penalty.}

\item{Zc}{a design matrix for those group-lasso-penalty free predictors. Can be missing. Default value is NULL.}

\item{k}{\eqn{k} is the size for each group in \eqn{Z}. All groups are with the same size. Number of groups is
\eqn{p = p1 / k }.}

\item{W}{a vector in length of p (the total number of groups), matrix with dimension \code{p1*p1} or character specifying function
used to caculate weight matrix for each group.
\itemize{
\item If vector, works as penalty factor. Separate penalty weights can be applied to each group of beta'ss to allow differential shrinkage. Can be 0 for some groups, which implies no shrinkage, and results in that group always being included in the model.
\item If matrix, a block diagonal matrix. Diagonal elements are inverted weights matrics for each group.
\item if character, user should provide the function.
}
Default value is rep(1, times = p).}

\item{intercept}{whether to include intercept. Default is TRUE.}

\item{A, b}{linear equality constraints \eqn{A\beta_p1 = b}, where \eqn{b} is a vector with length \eqn{k},
and \eqn{A} is a \eqn{k \times p1}{k*p1} matrix. Default values, \eqn{b} is a vector of 0's and
\code{A = kronecker(matrix(1, ncol = p), diag(k))}.}

\item{u}{\code{u} is the inital value for penalty parameter of augmented Lanrange method adopted in
outer loop - default value is 1.}

\item{mu_ratio}{\code{mu_ratio} is the increasing ratio for \code{u}. Default value is 1.01.
Inital values for scaled Lagrange multipliers are set as 0's.
If \code{mu_ratio} < 1,
there is no linear constraints included. Group lasso coefficients are estimated.}

\item{lam}{a user supplied lambda sequence. Typically, by leaving this option unspecified users can have the
program compute its own \code{lam} sequence based on \code{nlam} and \code{lambda.factor}.
Supplying a value of lambda overrides this.
If \code{lam} is provided but a scaler and \code{nlam}\eqn{>}1,
\code{lam} sequence is also created starting from \code{lam}.
If a sequence of lambda is provided, it is better to supply a decreasing one,
if not, the program will sort user-defined \code{lambda} sequence in decreasing order
automatically.}

\item{nlam}{the length of \code{lam} sequence. Default is 100.}

\item{lambda.factor}{the factor for getting the minimal lambda in \code{lam} sequence, where
\code{min(lam)} = \code{lambda.factor} * \code{max(lam)}.
\code{max(lam)} is the smallest value of \code{lam} for which all penalized group are zero's.
The default depends on the relationship between \eqn{n}
and \eqn{p1}
(the number of predictors to be penalized).
If \eqn{n >= p1}, the default is \code{0.001}, close to zero.
If \eqn{n < p1}, the default is \code{0.05}. A very small value of
\code{lambda.factor}
will lead to a saturated fit. It takes no effect if there is user-defined lambda sequence.}

\item{dfmax}{limit the maximum number of groups in the model. Useful for very large \eqn{p},
if a partial path is desired. Default is \eqn{p}.}

\item{pfmax}{limit the maximum number of groups ever to be nonzero. For example once a group enters the
model along the path, no matter how many times it exits or re-enters model through the path,
it will be counted only once. Default is \code{min(dfmax*1.5, p)}.}

\item{tol}{tolerance for vectors betas to be considered as none zero's. For example, coefficient
\eqn{\beta_j} for group j, if \eqn{max(abs(\beta_j))} < \code{tol}, set \eqn{\beta_j} as 0's.
Default value is 0.}

\item{outer_maxiter, outer_eps}{\code{outer_maxiter} is the maximun munber of loops allowed for Augmented Lanrange method;
and \code{outer_eps} is the convergence termination tolerance.}

\item{inner_maxiter, inner_eps}{\code{inner_maxiter} is the maximun munber of loops allowed for blockwise-GMD;
and \code{inner_eps} is the convergence termination tolerance.}
}
\value{
A list
\item{lam}{the actual sequence of lambda values used }
\item{df}{the number of nonzero groups in estimated coefficients for \code{Z} at each value of lambda}
\item{path}{a matrix of coefficients}
}
\description{
Fits regularization paths for linear constraints group-lasso penalized learning problems at a sequence of regularization parameters lambda.
}
