% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{coef.compCL}
\alias{coef.compCL}
\title{get coefficients or make coefficient predictions from and "compCL" object}
\usage{
\method{coef}{compCL}(object, s = NULL, ...)
}
\arguments{
\item{object}{fitted \code{\link{compCL}} object.}

\item{s}{value(s) of the penalty parameter \code{lam} at which predictions are required. Default is the
entire sequence used to create the model.}

\item{\dots}{not used.}
}
\value{
The coefficients at the requested values for \code{s}.
}
\description{
Computes the coefficients at the requested values for \code{lam} from a fitted
         \code{\link{compCL}} object.
}
\details{
\code{s} is the new vector at which predictions are requested. If \code{s} is not in the
         lambda sequence used for fitting the model, the \code{coef} function will use linear
         interpolation to make predictions. The new values are interpolated using a fraction of
         coefficients from both left and right \code{lam} indices.
}
\seealso{
\code{\link{compCL}}
}
