% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{coef.cv.compCL}
\alias{coef.cv.compCL}
\title{get coefficients from a cross-validation "cv.compCL" object.}
\usage{
\method{coef}{cv.compCL}(object, trim = FALSE, s = c("lam.min", "lam.1se"),
  ...)
}
\arguments{
\item{object}{fitted \code{\link{cv.compCL}} object.}

\item{trim}{logical, used the trimmed result or not. Default is FASLE.}

\item{s}{value(s) of the penalty parameter \code{lam} at which predictions are required.
Default is the value \code{s="lam.min"} stored on the CV \code{object}, it is the
optimal value of \code{lam} that gives minimum.
Alternatively \code{s="lambda.min"} can be used,  it is the largest value of \code{lam}
such that error is within 1 standard error of the minimum.
If \code{s} is numeric, it is taken as the value(s) of \code{lam} to be used.}

\item{\dots}{not used. Other arguments to predict.}
}
\value{
The coefficients at the requested values for \code{lam}.
}
\description{
This function gets coefficients or makes coefficient predictions from a cross-validated
\code{compCL} model, using the stored \code{"compCL.fit"} object,
and the optimal value chosen for \code{lam}.
}
\seealso{
\code{\link{cv.compCL}}
}
