% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{coef.cv.FuncompCGL}
\alias{coef.cv.FuncompCGL}
\title{get coefficients from a cross-validation "cv.FuncompCGL" object.}
\usage{
\method{coef}{cv.FuncompCGL}(object, trim = FALSE, s = c("lam.min",
  "lam.1se"), k = NULL, ...)
}
\arguments{
\item{object}{fitted \code{\link{cv.FuncompCGL}} object.}

\item{trim}{logical, used the trimmed result or not.}

\item{s}{value(s) of the penalty parameter \code{lam} at which predictions are required.
Default is the value \code{s="lam.min"} stored on the CV \code{object}, it is the
value of \code{lam} for optimal \code{k} such that gives minimum mean cross-validated error.
If \code{s} is numeric, it is taken as the value(s) of \code{lam} to be used.}

\item{k}{value of df of basis at which predictions are required. Default is \code{k = "k.min"},
it is the value of \code{k} that gives minimum mean cross-validated error. If \code{k}
is a scaler, it is taken as the value of \code{k} to be used and it must be stored in
\code{\link{cv.FuncompCGL}} object.}

\item{\dots}{not used.}
}
\value{
The coefficients at the requested values for \code{lam} and \code{k}.
}
\description{
This function gets coefficients or makes coefficient predictions from a cross-validated
\code{FuncompCGL} model, using the stored \code{"FuncompCGL.fit"} object,
and the optimal value chosen for \code{lam} and\code{k} (degree of freedom of basis).
Questionable!!!
}
\seealso{
\code{\link{cv.FuncompCGL}}
}
