% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{comp_Model}
\alias{comp_Model}
\title{Simulation for composition data}
\usage{
comp_Model(n, p, rho = 0.2, sigma = 0.5, gamma = 0.5, add.on = 1:5,
  beta = c(c(1, -0.8, 0.6, 0, 0, -1.5, -0.5, 1.2), rep(0, times = p - 8)),
  beta0 = 1, intercept = TRUE)
}
\arguments{
\item{n}{sample size.}

\item{p}{size of compositional predictors falling in \eqn{S^p}.}

\item{rho}{a scaler used to generate \code{p*p} autocorrelation matrix.}

\item{sigma}{standard deviation for noise terms, which follows iid norm distribution with mean 0.}

\item{gamma, add.on}{To generate compsotional data, we first generate \code{n*p} matrix \code{X} from multivariate normal distribution and
than transform it to compostional data.
\code{gamma} is a scaler and add.on is a index vector. Assign \code{log(p)*gamma} to means of \code{X} for columns
\code{add.on}, while others with 0's.}

\item{beta}{coefficients for composition variables.}

\item{beta0}{coefficient for intercept.}

\item{intercept}{including intercept or not to generate response variable. Default value is TRUE.}
}
\value{
A list of data
\item{y}{A vector of simulated response}
\item{Z}{A matrix of simulated log of compositional predictors, with size n*p}
\item{Zc}{A matrix of simulated covariates except compositional ones}
\item{intercept}{A boolean indicating whether including intercept}
\item{beta}{true coefficient beta}
\item{X.comp}{A matrix of simulated compositional predictors of size n*p}
}
\description{
Simulate compositional data.
}
\details{
The setup of this simulation follows Variable selection in regression with compositional covariates by WEI LIN, PIXU SHI, RUI FENG AND HONGZHE LI.
}
\examples{

n = 50
p = 30
rho = 0.2
sigma = 0.5
gamma = 0.5
add.on = 1:5
beta = c(1, -0.8, 0.6, 0, 0, -1.5, -0.5, 1.2)
beta = c( beta, rep(0, times = p - length(beta)) )
intercept = FALSE
Comp_data = comp_Model(n = n, p = p, rho = rho, sigma = sigma, gamma  = gamma,
                       add.on = add.on, beta = beta, intercept = intercept)

}
