% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compreg.r
\name{compreg}
\alias{compreg}
\title{Run a Compositional Regression on All Profiles}
\usage{
compreg(
  form,
  composition,
  data,
  count,
  total,
  relative,
  n_min,
  transform = c("alr", "clr", "ilr", "apt", "cpt", "ipt"),
  ...
)
}
\arguments{
\item{form}{A right-hand sided formula of the independent variables in the regression model.}

\item{composition}{A \code{tidy-select} statement that identifies the components of the composition in the data.  Note,
the data must be in wide-format already.}

\item{data}{A data frame that contains variables in \code{formula} and \code{composition}.}

\item{count}{Logical indicating whether the data represent counts or not.}

\item{total}{Logical indicating whether the components of the composition convey information about the total (\code{TRUE}) or
their sum is irrelevant or uninformative (\code{FALSE}).}

\item{relative}{Logical indicating whether the components of the composition are ratio level (\code{TRUE}) or interval level (\code{FALSE}).}

\item{n_min}{Number of observations required to estimate a model for the profile.}

\item{transform}{Transformation used for the composition.  Options are \code{"alr"}, \code{"clr"}, \code{"ilr"}, \code{"apt"}, \code{"cpt"} and \code{"ipt"}.
For more information, see \code{help(package="compositions")}.}

\item{...}{Other arguments passed down to \code{lm()}.}
}
\description{
Runs a compositional regression model on all profiles with at least \code{n_min} observations
and saves the models in a way that is amenable to the post-model estimation routines in the
package.
}
\details{
For more information on the \code{count}, \code{total} and \code{relative} arguments, see \code{system.file("doc/compositions_v2.html", package="compositions")}
}
