% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_ErrorFun.R
\name{ERROR_fun}
\alias{ERROR_fun}
\title{error function}
\usage{
ERROR_fun(beta_fit, beta_true, basis_fit, basis_true, sseq, m, p, Nzero_group,
  tol = 0)
}
\arguments{
\item{beta_fit}{fitted coefficients vector}

\item{beta_true}{true coefficientse vector}

\item{basis_fit}{basis matrix with selected df \code{k} and \code{basis_fun}}

\item{basis_true}{basis matrix with true df \code{k} and \code{basis_fun}}

\item{sseq}{sequence of time points used to generate basis matrix}

\item{m}{number of time-invariant variables}

\item{p}{number of compositional varialbes}

\item{Nzero_group}{number of true None zero group}

\item{tol}{tolerance}
}
\value{
a list
\item{Non.zero}{None zero rows in coefficient matrix \code{beta_C}. }
\item{class_error}{classification errors, including FP, FN, FPR, FNR, etc.}
\item{coef_error}{\itemize{
                           \item \code{L2.cont} L2 norm for time-invariant coefficients error
                           \item \code{L2.comp, L2_inf.comp, L2_L1.comp} if selected df \code{k} is
                                 the same as true degree freedom of basis, coefficients matrix error is
                                 calculated}
                 }
\item{curve_error}{norms for curve error by function norms}
\item{group_norm}{L2 norm for rows in coefficients matrix \code{beta_C}}
}
\description{
calculate classification, prediction and accuracy errors for fitted
\code{FuncompCGL} model in simulation.
}
