% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.r
\name{fd}
\alias{fd}
\title{First Differences for Compositional Regression Models}
\usage{
fd(models, variable, diff, values = NULL, R = 2500, ...)
}
\arguments{
\item{models}{A list of models estimated by \code{compreg}.}

\item{variable}{A string identifying the name of the variable whose first difference will be calculated.}

\item{diff}{A string identifying the kind of first difference desired - \code{"sd"} for standard deviation change
around observed values, \code{"range"} for a maximal change for the variable or \code{"unit"} for a one-unit change around
observed values.}

\item{values}{An optional vector of values to use for the first difference.  If \code{values} is specified, it overrides
\code{diff}.}

\item{R}{Number of bootstrap samples to use.}

\item{...}{Not implemented}
}
\description{
Thanks to the non-linear transformation of the composition, the effects are on each component of the composition
are not as straightforward to assess as the linear model may suggest.  We use first differences to evaluate
the effect of the variables in the model.
}
\details{
This function does a parametric boostrap of the coefficient matrix to estimate the variability of the
predicted compositions.  This permits the calculation of confidence intervals for first differences.
}
