% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_ErrorFun.R
\name{Genbasis}
\alias{Genbasis}
\title{Basis matrix generation}
\usage{
Genbasis(sseq, df, degree, type = c("bs", "fourier", "OBasis"))
}
\arguments{
\item{sseq}{the predictor variable.}

\item{df}{degree of freedom.}

\item{degree}{degree of piecewise polynomial - default is 3. Used in \code{bs} and \code{OBasis}.}

\item{type}{type of basis}
}
\value{
basis matrix of size \code{length(sseq)} by \code{df}
}
\description{
generate the basis matrix for polynomial spline, fourier, and orthogonalsplinebasis B-splines.
}
\details{
For type \code{bs} and \code{OBasis}, \code{intercept} is set to \code{TRUE}, minimum value for
\code{df} is 4.
For type \code{fourier}, \code{df} should be a even number. If not, \code{df} is automatically
set to \code{df} - 1.
}
