% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIC.R
\name{GIC.compCL}
\alias{GIC.compCL}
\title{GIC cirterion selection for compCL}
\usage{
GIC.compCL(y, Z, Zc = NULL, intercept = FALSE, lam = NULL, ...)
}
\arguments{
\item{y}{a vector of response variable with length n.}

\item{Z}{a \eqn{n*p} matrix of compositional data or categorical data.
If \code{Z} is categorical data, i.e., row sums of \code{Z} differ from 1, \code{Z} is transformed to compositional data
by diving each row of it's corresponding sum.
\code{Z} could NOT include entry of 0.}

\item{Zc}{a \eqn{n*p_c} design matrix of other covariates considered except composition . Default is \code{NULL}.}

\item{intercept}{Whether to include intercept in the model. Default is TRUE.}

\item{lam}{a user supplied lambda sequence. Typically, by leaving this option unspecified users can have the
program compute its own \code{lam} sequence based on \code{nlam} and \code{lambda.factor}.
Supplying a value of lambda overrides this.
If \code{lam} is provided but a scaler and \code{nlam}\eqn{>}1,
\code{lam} sequence is also created starting from \code{lam}.
If a sequence of lambda is provided, it is better to supply a decreasing one,
if not, the program will sort user-defined \code{lambda} sequence in decreasing order
automatically.}

\item{\dots}{other arguments that can be passed to compCL.}
}
\value{
an object of class \code{\link{GIC.compCL}} is returned.
\item{compCL.fit}{a fitted \code{\link{compCL}} object for the full data}
\item{lam}{the values of \code{lam} used in the fits}
\item{GIC}{a vector of GIC values for each \code{lam}}
\item{lam.min}{\code{lam} value such that minimize \eqn{GIC(\lambda)} }
}
\description{
Calculate GIC for compCL, return value of \code{lam}.
The function follows Variable selection in regression with compositional covariates by
WEI LIN, PIXU SHI, RUI FENG AND HONGZHE LI
}
\details{
\deqn{\textrm{GIC}(\lambda) = \log{\hat{\sigma}^2_\lambda}
                      + (s_\lambda - 1) \frac{\log{\log{n}}}{n} \log{max(p, n)} },
where \eqn{\hat{\sigma}^2_\lambda} is the MSE for fitted path.


\deqn{F_{n} =\frac{\phi^{n} - \psi^{n}}{\sqrt{5}}.}
}
\examples{

p = 30
n = 50
beta = c(1, -0.8, 0.6, 0, 0, -1.5, -0.5, 1.2)
beta = c(beta, rep(0, times = p - length(beta)))
Comp_data = comp_Model(n = n, p = p,
                            rho = 0.2, sigma = 0.5,
                            gamma  = 0.5, add.on = 1:5,
                            beta = beta, intercept = FALSE)
Comp_data$Zc
GICm <- GIC.compCL(y = Comp_data$y,
                   Z = Comp_data$X.comp, Zc = Comp_data$Zc,
                   intercept = Comp_data$intercept,
                   lam = NULL,lambda.factor = 0.0001,
                   dfmax = p, outer_eps = 1e-10, mu_ratio = 1)
coef(GICm)
plot(GICm)
}
