% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_profiles.r
\name{id_profiles}
\alias{id_profiles}
\title{Identify Choice Profiles}
\usage{
id_profiles(x, sel = everything(), ...)
}
\arguments{
\item{x}{A matrix or data frame that represents the elements of a composition. Components with values 0 or \code{NA}
will be considered as absent from that observation's composition.}

\item{sel}{A \code{tidy-select} statement that identifies the elements of the composition.  Note, the data must already be
in wide format.  The default is \code{everything()} which chooses all columns.}

\item{...}{Not currently implemented.}
}
\description{
Variables that comprise a composition may have cases where not all potential components of the composition
are observed.   We refer to the set of observed components of the composition as a "profile".  This function
identifies the existing profiles and gives a count for each one.
}
