% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral.R
\name{ITG_trap}
\alias{ITG_trap}
\title{Itegration by trapezoidal}
\usage{
ITG_trap(X, basis, sseq, T.name = "TIME", from = 0, to = 1)
}
\arguments{
\item{X}{a data.frame or vector
\itemize{
\item If X is a data.frame, dim(X)=nobs*(p+1), nobs is the total number of observations, p is size of covariates
\item If X is a vector, X is a sequence of observed time points}}

\item{basis}{a matrix of basis}

\item{sseq}{a sequence used as predictor variable to generate basis}

\item{T.name}{a character specifying name for time varaible in X}

\item{from, to}{two scalers defining domain for integration - defalt values from = 0, to = 1.}
}
\value{
a scaler approximating integral. If X is a vector, taking integration for basis on user provided time points sequence.
}
\description{
Sum up area under trapezoidal formulated by values of function at two adjacent observed time points.
}
