% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{KKT_glasso}
\alias{KKT_glasso}
\title{check KKT condition for solution for "FuncompCGL".
check KKT condition for solution for "FuncompCGL".}
\usage{
KKT_glasso(y, Z, p, k, A, b, lam, beta, tol = 1e-10)
}
\arguments{
\item{y}{response}

\item{Z}{column combination of compositional variables integration and control variables}

\item{p}{numbers of compositional variables}

\item{k}{degree of freedom for basis}

\item{A, b}{linear constraints}

\item{lam}{\code{lambda} used}

\item{beta}{coefficients vector}

\item{tol}{tolerance de}
}
\value{
\item{summary}{logicl, KKT condition satisfied or not}
\item{feasible}{feasible condition}
\item{primary}{primary condition}
\item{KKT_mutiplier}{KKT_mutiplier}
}
\description{
check KKT condition for solution for "FuncompCGL".
check KKT condition for solution for "FuncompCGL".
}
