% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.compCL}
\alias{plot.compCL}
\title{plot coefficients from a \code{"compCL"} object}
\usage{
\method{plot}{compCL}(x, xlab = c("norm", "log"), label = FALSE, ...)
}
\arguments{
\item{x}{fitted \code{"compCL"} model.}

\item{xlab}{What is on the X-axis. "\code{lambda} against \code{log(lambda)} (default)
or "\code{norm}" against \code{L1-norm} .}

\item{label}{if TRUE, lable the curve with variable sequence numbers.}

\item{\dots}{Other graphical parameters to plot.}
}
\description{
Produces a coefficient profile plot of the coefficient paths for a fitted
\code{"compCL"} object.
}
\details{
A plot is prodoced, and nothing is returned.
}
\seealso{
\code{\link{compCL}}
}
