% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.cv.compCL}
\alias{plot.cv.compCL}
\title{plot the cross-validation curve produced by "cv.compCL" object.}
\usage{
\method{plot}{cv.compCL}(x, xlab = c("log", "-log", "lambda"), ...)
}
\arguments{
\item{x}{fitted \code{"cv.compCL"} object.}

\item{xlab}{Either plot against \code{log(lambda)} (default) or \code{lambda} or \code{-log(lambda)}.}

\item{\dots}{Other graphical parameters to plot.}
}
\description{
Plots the cross-validation curve, and upper and lower standard deviation
curves, as a function of the \code{lambda} values used.
}
\details{
A plot is produced, and nothing is returned.
}
\seealso{
\code{\link{cv.compCL}}
}
