% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.cv.FuncompCGL}
\alias{plot.cv.FuncompCGL}
\title{plot the cross-validation curve produced by cv.FuncompCGL}
\usage{
\method{plot}{cv.FuncompCGL}(x, xlab = c("log", "lambda"), trim = FALSE,
  k_list, ...)
}
\arguments{
\item{x}{fitted \code{"cv.FuncompCGL"}.}

\item{xlab}{Either plot against \code{log(lambda)} (default) or \code{lambda}.}

\item{trim}{Logic value, determining use trimmed result stored in \code{cv.FumcompCGL} or not. Default}

\item{k_list}{a vector, determining result of which \code{k} (degree freedom) to plot.
If not provided, cross-validation curve for k that associate with \code{lambda.min} (blue)
and \code{lambda.1se} (red) are plotted.}

\item{\dots}{Other graphical parameters to plot}
}
\description{
Plots the cross-validation curve for different degree of freedom \code{k},
and upper and lower standard deviation curves,
as a function of the \code{lambda} values used.
}
\details{
A plot is prodoced, and nothing is returned.
}
\seealso{
\code{\link{cv.FuncompCGL}}
}
