% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.FuncompCGL}
\alias{plot.FuncompCGL}
\title{plot coefficients from a \code{"FuncompCGL"} object}
\usage{
\method{plot}{FuncompCGL}(x, p, k, ylab = c("L2", "L1"), xlab = c("log",
  "lambda"), ...)
}
\arguments{
\item{x}{fitted \code{"FuncompCGL"} model.}

\item{p}{numbers of compositional variables.}

\item{k}{degree of freedom of basis used.}

\item{ylab}{What is on the Y-axis, L1-norm or L2-norm (default) of each group of coefficients.}

\item{xlab}{Either plot against \code{log(lambda)} (default) or \code{lambda}.}

\item{\dots}{Other graphical parameters to plot.}
}
\description{
Produces a coefficient profile plot of the coefficient paths for a fitted
\code{"FuncompCGL"} object.
}
\details{
A plot is prodoced, and nothing is returned.
}
\seealso{
\code{\link{FuncompCGL}}
}
