% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.GIC.compCL}
\alias{plot.GIC.compCL}
\title{plot the GIC curve produced by "GIC.compCL" object.}
\usage{
\method{plot}{GIC.compCL}(x, xlab = c("log", "-log", "lambda"), ...)
}
\arguments{
\item{x}{fitted \code{"GIC.compCL"} object.}

\item{xlab}{Either plot against \code{log(lambda)} (default) or \code{-log{lambda}} or \code{lambda}.}

\item{\dots}{Other graphical parameters to plot.}
}
\description{
Plots the CIC curve as a function of the \code{lambda} values used.
}
\details{
A plot is produced, and nothing is returned.
}
\seealso{
\code{\link{GIC.compCL}}
}
