% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.GIC.FuncompCGL}
\alias{plot.GIC.FuncompCGL}
\title{plot the GIC curve produced by "GIC.FuncompCGL" object.}
\usage{
\method{plot}{GIC.FuncompCGL}(x, xlab = c("log", "-log", "lambda"), k_list,
  ...)
}
\arguments{
\item{x}{fitted \code{"GIC.FuncompCGL"} object.}

\item{xlab}{Either plot against \code{log(lambda)} (default) or \code{-log{lambda}} or \code{lambda}.}

\item{k_list}{a vector, determining result of which \code{k} (degree freedom) to plot.
If not provided, GIC curve for k that associate with \code{lambda.min} (red)
is plotted.}

\item{\dots}{Other graphical parameters to plot.}
}
\description{
Plots the CIC curve as a function of the \code{lambda} values used.
}
\details{
A plot is produced, and nothing is returned.
}
