% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{predict.compCL}
\alias{predict.compCL}
\title{make predictions from a "compCL" object}
\usage{
\method{predict}{compCL}(object, newx, newzc = NULL, s = NULL, ...)
}
\arguments{
\item{object}{fitted \code{\link{compCL}} object.}

\item{newx, newzc}{newx is the matrix of new values for log composition and newzc is the matrix for confounding
matrix. Joint of newx and newzc is the design matrix at which predictions are to be made.
Default value for newzc is \code{NULL}.}

\item{s}{value(s) of the penalty parameter \code{lam} at which predictions are required.
Default is the entire sequence used to create the model.}

\item{\dots}{Not used. Other arguments to predict.}
}
\value{
prediction values at the requested values for \code{s}.
}
\description{
predicts fitted values and class labels from a fitted \code{\link{compCL}} object.
}
\details{
\code{s} is the new vector at which predictions are requested. If \code{s} is not in the lambda
       sequence used for fitting the model, the \code{predict} function will use linear interpolation
       to make predictions. The new values are interpolated using a fraction of predicted values from
       both left and right \code{lam} indices.
}
\seealso{
\code{\link{compCL}}
}
