% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{predict.cv.compCL}
\alias{predict.cv.compCL}
\title{make predictions from a \code{"cv.compCL"} object.}
\usage{
\method{predict}{cv.compCL}(object, Znew, Zcnew = NULL, s = c("lam.min",
  "lam.1se"), trim = FALSE, ...)
}
\arguments{
\item{object}{fitted \code{"cv.compCL"} model}

\item{Znew}{new compositional data}

\item{Zcnew}{new time-invariant covarites}

\item{s}{\code{"lam.min"} or \code{"lam.1se"}}

\item{trim}{\code{"FALSE"}}

\item{\dots}{not used. Other arguments to predict.}
}
\description{
This function makes prediction from a cross-validated \code{compCL} model,
using the stored \code{compCL.fit} object and the optimal value chose for \code{lambda}.
}
\examples{
p = 30
n = 50
beta = c(1, -0.8, 0.6, 0, 0, -1.5, -0.5, 1.2)
beta = c( beta, rep(0, times = p - length(beta)) )
Comp_data = comp_Model(n = n, p = p,
                       rho = 0.2, sigma = 0.5,
                       gamma  = 0.5, add.on = 1:5,
                       beta = beta, intercept = FALSE)
test_data = comp_Model(n = 100, p = p,
                       rho = 0.2, sigma = 0.5,
                       gamma  = 0.5, add.on = 1:5,
                       beta = beta, intercept = FALSE)
cvm <- cv.compCL(y = Comp_data$y,
                 Z = Comp_data$X.comp, Zc = Comp_data$Zc,
                 intercept = Comp_data$intercept,
                 lam = NULL, nfolds = 10, trim = 0.05, lambda.factor = 0.0001,
                 dfmax = p, mu_ratio = 1, outer_eps = 1e-10, inner_eps = 1e-8, inner_maxiter = 1e4)
y_hat = predict(cvm, Znew = test_data$X.comp, Zcnew = test_data$Zc)
plot(test_data$y, y_hat)
abline(a = 0, b = 1, col = "red")

}
\seealso{
\code{\link{cv.compCL}}
}
