% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{predict.GIC.compCL}
\alias{predict.GIC.compCL}
\title{make predictions from a \code{"GIC.compCL"} object.}
\usage{
\method{predict}{GIC.compCL}(object, Znew, Zcnew = NULL, s = "lam.min", ...)
}
\arguments{
\item{object}{fitted \code{"GIC.compCL"} model}

\item{Znew}{new compositional data}

\item{Zcnew}{new time-invariant covarites}

\item{s}{\code{"lam.min"} or user provided value}

\item{\dots}{not used. Other arguments to predict.}
}
\description{
This function makes prediction from a GIC  \code{compCL} model,
using the stored \code{compCL.fit} object and the optimal value chose for \code{lambda}.
}
\examples{
p = 30
n = 50
beta = c(1, -0.8, 0.6, 0, 0, -1.5, -0.5, 1.2)
beta = c(beta, rep(0, times = p - length(beta)))
Comp_data = comp_Model(n = n, p = p,
                            rho = 0.2, sigma = 0.5,
                            gamma  = 0.5, add.on = 1:5,
                            beta = beta, intercept = FALSE)
test_data = comp_Model(n = 100, p = p,
                            rho = 0.2, sigma = 0.5,
                            gamma  = 0.5, add.on = 1:5,
                            beta = beta, intercept = FALSE)
GICm <- GIC.compCL(y = Comp_data$y,
                   Z = Comp_data$X.comp, Zc = Comp_data$Zc,
                   intercept = Comp_data$intercept,
                   lam = NULL,lambda.factor = 0.0001,
                   dfmax = p, outer_eps = 1e-10, mu_ratio = 1)
y_hat = predict(GICm, Znew = test_data$X.comp, Zcnew = test_data$Zc)
plot(y_hat, test_data$y)
abline(a = 0, b = 1, col = "red")
}
\seealso{
\code{\link{GIC.compCL}}
}
