% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print.FuncompCGL}
\alias{print.FuncompCGL}
\title{print a "FuncompCGL" object}
\usage{
\method{print}{FuncompCGL}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fitted \code{\link{FuncompCGL}} object.}

\item{digits}{significant digits in printout.}

\item{\dots}{Not used. Other arguments to predict.}
}
\value{
a two-column matrix, the first columns is the number of nonzero group counts and
        the second column is Lambda.
}
\description{
Print the nonzero counts for composition varaibles at each lambda along the FuncompCGL path.
}
\seealso{
\code{\link{FuncompCGL}}
}
